+++ app/Notifications/Donor/SendPaymentReceipt.php
@@ -43,11 +43,15 @@ class SendPaymentReceipt extends Notification
         $general = App\General::get();
         $cp_email[0] = $general->where('key', 'Contact Person Email 1')->first()->value;
         $cp_email[1] = $general->where('key', 'Contact Person Email 2')->first()->value;
+
         $url = 'hello@seedscholarship.org';
         return (new MailMessage)
             ->from($url, 'SEED Scholarship')
             ->subject("Terimakasih Atas Donasinya | Receipt #{$this->data->donor->donorTransactions[0]->invoice_no}")
-            ->markdown('email.DonorPaymentReceipt', ['data' => $this->data, 'cp_email' => $cp_email])
+            ->markdown('email.DonorPaymentReceipt', [
+                'data' => $this->data,
+                'cp_email' => $cp_email,
+            ])
             ->attach(storage_path('app') . "/transaction/{$this->data->period->year}/{$this->data->donor->id}/{$this->data->donor->donorTransactions[0]->id}/invoice/{$this->data->donor->donorTransactions[0]->invoice_no}.pdf");
 
     }